feature: User情報をグローバル変数で管理
